refactor(inference): change the structure of the inference module
